feat: moved game info caches to redis
chore: renamed game/gameCache to game/cacheInterface for clarity
