+++ cspell.json
@@ -1,6 +1,17 @@
 {
    "version": "0.2",
-   "words": ["reversi", "kolb", "unsub", "preconfigured"],
+   "words": [
+      "reversi",
+      "kolb",
+      "unsub",
+      "preconfigured",
+      "hmget",
+      "lrange",
+      "hdel",
+      "lrem",
+      "lpush",
+      "hset"
+   ],
    "ignorePaths": ["package.json", "package-lock.json"],
    "useGitignore": true,
    "cache": {
