+++ package.json
@@ -1,6 +1,6 @@
 {
    "name": "reversi-next",
-   "version": "0.2.4",
+   "version": "0.3.0",
    "private": true,
    "scripts": {
       "dev": "next dev --turbopack",
