+++ src/lib/gameManager/__test__/join.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/consistent-type-assertions */
 import { join } from '../join';
-import { getGame } from '@/lib/game/gameCache';
+import { getGame } from '@/lib/game/cacheInterface';
 import type { Client } from '@/lib/client/Client';
 import { upgradeGame } from '@/lib/gameManager/upgradeGame';
 
-jest.mock('@/lib/game/gameCache', () => ({
+jest.mock('@/lib/game/cacheInterface', () => ({
    getGame: jest.fn(),
 }));
